requirements.txt
anyascii==0.1.7
attrs==19.1.0
beautifulsoup4==4.9.3
blis==0.7.4
boto==2.49.0
boto3==1.9.131
botocore==1.12.131
bs4==0.0.1
bz2file==0.98
catalogue==1.0.0
certifi==2019.3.9
cffi==1.14.5
chardet==3.0.4
click==7.1.2
clickclick==20.10.2
connexion==2.7.0
contractions==0.0.48
cymem==2.0.5
decorator==4.4.0
Deprecated==1.2.12
docutils==0.14
en-core-web-lg==2.3.1
Flask==1.1.2
gensim==3.8.3
glob2==0.7
htmlmin==0.1.12
idna==2.8
inflection==0.5.1
ipython-genutils==0.2.0
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.9.4
joblib==1.0.1
jsonschema==3.2.0
jupyter-core==4.4.0
lxml==4.6.3
MarkupSafe==1.1.1
murmurhash==1.0.5
nltk==3.5
numpy==1.20.1
openapi-schema-validator==0.1.4
openapi-spec-validator==0.3.0
plac==1.1.3
preshed==3.0.5
pyahocorasick==1.4.1
pycodestyle==2.7.0
pycparser==2.20
pymisp==2.4.140
pymongo==3.7.2
pyrsistent==0.14.11
python-dateutil==2.8.1
PyYAML==5.4.1
regex==2020.11.13
requests==2.25.1
rope==0.18.0
s3transfer==0.2.0
scipy==1.6.1
six==1.12.0
smart-open==1.8.2
soupsieve==2.2.1
spacy==2.3.5
srsly==1.0.5
swagger-ui-bundle==0.0.8
textsearch==0.0.21
thinc==7.4.5
threadpoolctl==2.1.0
tqdm==4.59.0
traitlets==4.3.2
urllib3==1.24.2
waitress==2.0.0
wasabi==0.8.2
watchdog==2.0.2
Werkzeug==1.0.1
wrapt==1.12.1
